Upgrade to Express 4 routing style with mountable middleware app. Instead of `require('./config')` everywhere, pass the configuration through.
