+++ package.json
@@ -8,12 +8,20 @@
     "url": "git://github.com/andzdroid/mongo-express.git"
   },
   "dependencies": {
-    "express": "~3.4.8",
-    "mongodb": "~1.3.23",
+    "async": "~0.8.0",
+    "basic-auth-connect": "^1.0.0",
+    "body-parser": "^1.12.4",
     "consolidate": "~0.10.0",
+    "cookie-parser": "^1.3.5",
+    "errorhandler": "^1.3.6",
+    "express": "^4.12.4",
+    "express-session": "^1.11.2",
+    "method-override": "^2.3.3",
+    "mongodb": "~1.3.23",
+    "morgan": "^1.5.3",
+    "serve-favicon": "^2.2.1",
     "swig": "~1.3.2",
-    "underscore": "~1.6.0",
-    "async": "~0.8.0"
+    "underscore": "~1.6.0"
   },
   "devDependencies": {
     "mocha": "~1.18.2",
